exam-slammr-rest/exam-slammr-rest-common/src/main/java/uk/co/claritysoftware/exam/slammr/web/interceptor/ClearMdcInterceptor.java
package uk.co.claritysoftware.exam.slammr.web.interceptor;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import org.slf4j.MDC;
import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;

/**
 * Spring web interceptor to clear the MDC context on request completion
 */
public class ClearMdcInterceptor extends HandlerInterceptorAdapter {

    @Override
    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) {
        MDC.clear();
    }
}
